+++ frontend/javascripts/admin/task/task_create_form_view.tsx
@@ -34,7 +34,7 @@ import { AsyncButton } from "components/async_clickables";
 import { formatDateInLocalTimeZone } from "components/formatted_date";
 import SelectExperienceDomain from "components/select_experience_domain";
 import { saveAs } from "file-saver";
-import { unwrapOrThrow } from "libs/api_result";
+import { unwrapOrThrow } from "admin/api/api_result";
 import { coalesce, pluralize, tryToAwaitPromise } from "libs/utils";
 import { Vector3Input, Vector6Input } from "libs/vector_input";
 import isEqual from "lodash-es/isEqual";
