+++ frontend/javascripts/components/secured_route.tsx
@@ -5,7 +5,7 @@ import {
 } from "admin/organization/pricing_plan_utils";
 import { getUnversionedAnnotationInformation } from "admin/rest_api";
 import { PageUnavailableForYourPlanView } from "components/pricing_enforcers";
-import { unwrapOrThrow } from "libs/api_result";
+import { unwrapOrThrow } from "admin/api/api_result";
 import { useWkSelector } from "libs/react_hooks";
 import { isUserAdminOrManager } from "libs/utils";
 import type React from "react";
