+++ frontend/javascripts/viewer/view/ai_jobs/hooks/fetch_annotation_infos.ts
@@ -5,7 +5,7 @@ import {
   getTracingForAnnotationType,
   getUnversionedAnnotationInformation,
 } from "admin/rest_api";
-import { unwrapOrThrow } from "libs/api_result";
+import { unwrapOrThrow } from "admin/api/api_result";
 import Toast from "libs/toast";
 import { computeBoundingBoxFromBoundingBoxObject, point3ToVector3 } from "libs/utils";
 import { AnnotationLayerEnum, type APIAnnotation, type ServerVolumeTracing } from "types/api_types";
